delete individuals when they die.
or rather, mark them as being ready for deletion, which will actually happen if someone else replaces them in the hashtable, or if the hashtable extends its capacity.
And everything gets deleted on shutdown.

In order to delete individuals right away, we'd neet to blank out their reference in hashtable.remove.
It's not clear that this would be worth while though.