adjust enemy difficulty.
monsters have a narrower range of XP levels they can spawn at.
this makes it easier to predict how strong an enemy is by looking at their species.

the difficulty of monsters present at map generation time is only determined by the level depth, not your XP level.
each level has a 'miniboss' monster, which is always the same for a given level depth,
and is 1 level harder than normal monsters for that level.
this makes level depths more recognizable and predictable, e.g. level 3 is the one with the scorpion miniboss.

when levels are generated, the initially present monsters have a random XP level from 0 to the dungeon level,
biased towards being more difficult.
this makes it so that you feel stronger when you blow through weak monsters you had trouble with earlier in the game
(at least that's the idea.).

adjusted the hitpoints and damage of several monsters.
blobs are now easy monsters instead of difficult monsters.