perceived positions use consistent coordinates
it's no longer true that self.rel_position is always 0,0.
you still don't get true absolute coordinates,
but you get what looks like absolute coordinates.
every individual is told that the absolute coordinates origin is their spawn location.

in the future, if we want to have a discontinuity in coordinates,
like falling unconscious while being moved around, or experiencing non-euclidean geometry,
that can be communicated as special messages from the server.
that will also prompt the client to add a page to the map layer collection,
which the player might presumably be able to shift around and position relative to other pages
to reconstruct a single continuity. to be implemented later(TM).